Added tests for list of strings to escaped environment var string.
