+++ Bonobo.Git.Server/Extensions/UserExtensions.cs
@@ -7,7 +7,7 @@ using System.Security.Principal;
 
 namespace Bonobo.Git.Server
 {
-    public static class UserExtensions
+    public static partial class UserExtensions
     {
         static string GetClaimValue(this IPrincipal user, string claimName)
         {
@@ -142,6 +142,36 @@ namespace Bonobo.Git.Server
 
             return string.Empty;
         }
+
+        // http://stackoverflow.com/questions/915745/thoughts-on-foreach-with-enumerable-range-vs-traditional-for-loop
+        public static IEnumerable<int> To(this int from, int to)
+        {
+            if (from < to)
+            {
+                while (from <= to)
+                {
+                    yield return from++;
+                }
+            }
+            else
+            {
+                while (from >= to)
+                {
+                    yield return from--;
+                }
+            }
+        }
+
+        public static IEnumerable<T> Step<T>(this IEnumerable<T> source, int step)
+        {
+            if (step == 0)
+            {
+                throw new ArgumentOutOfRangeException("step", "Param cannot be zero.");
+            }
+
+            return source.Where((x, i) => (i % step) == 0);
+        }
+
         public static string StringlistToEscapedStringForEnvVar(IEnumerable<string> items, string separator = ",")
         {
             var y = items.Select(x => x.Replace(@"\", @"\\").Replace(separator, @"\"+separator));
